send log to logstash

+++ src/main/kotlin/com/example/frontServer/dto/logstash/BoardWriteEventLog.kt
@@ -1,7 +1,7 @@
 package com.example.frontServer.dto.logstash
 
 data class BoardWriteEventLog(
-    val logEvent: Int,
+    val logEvent: Long,
     val boardId: Long,
     val userId: Long,
 )
